+++ tests/ChargeFormTest.php
@@ -2,7 +2,7 @@
 
 use idarex\pingppyii2\ChargeForm;
 
-class ChargeFormTest extends PHPUnit_Framework_TestCase
+class ChargeFormTest extends TestCase
 {
     public function testCreate()
     {
@@ -30,5 +30,40 @@ class ChargeFormTest extends PHPUnit_Framework_TestCase
 
         $signature = $chargeForm->getWechatSignature(null, 'https://m.idarex.com');
         $this->assertNotEmpty($signature, 'Signature can not be empty');
+    }
+
+    /**
+     * @depends testCreate
+     */
+    public function testCreateWithPrivateKeyPath()
+    {
+        $config = Yii::$app->getComponents()['pingpp'];
+        $config['privateKeyPath'] = dirname(__FILE__) . '/rsa_private_key.pem';
+        $component = Yii::createObject($config);
+
+        $chargeForm = new ChargeForm();
+        $chargeForm->component = $component;
+        $chargeForm->load(require 'data/charge.php', '');
+
+        $this->assertTrue($chargeForm->validate() && $chargeForm->create());
+        $this->assertTrue(is_array($chargeForm->getCharge(true)));
+    }
+
+    /**
+     * @depends testCreate
+     */
+    public function testCreateWithPrivateKey()
+    {
+        $config = Yii::$app->getComponents()['pingpp'];
+        $config['privateKey'] = file_get_contents(dirname(__FILE__) . '/rsa_private_key.pem');
+
+        $component = Yii::createObject($config);
+
+        $chargeForm = new ChargeForm();
+        $chargeForm->component = $component;
+        $chargeForm->load(require 'data/charge.php', '');
+
+        $this->assertTrue($chargeForm->validate() && $chargeForm->create());
+        $this->assertTrue(is_array($chargeForm->getCharge(true)));
     }
 }