+++ tests/data/transfer.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+    'amount' => 100,
+    'order_no' => '20160419' . (string)time(),
+    'currency' => 'cny',
+    'channel' => 'wx_pub',
+    'type' => 'b2c',
+    'recipient' => 'o9zpMs9jIaLynQY9N6yxcZ',
+    'description' => 'testing',
+    'user_name' => 'User Name',
+    'force_check' => true,
+];